+++ commentsCounter.test.js
@@ -1,4 +1,3 @@
-// eslint-disable-next-line import/no-extraneous-dependencies
 import { describe, test, expect } from '@jest/globals';
 import countComments from './src/modules/commentsCounter.js';
