+++ src/modules/likesCounter.js
@@ -1,4 +1,3 @@
-// Returns the length of the node (the divs with the series class).
 const likesCounter = () => {
   if (!document.querySelectorAll('.item')) {
     return 0;
